Prepare to support different PLONKY2 target types
The new possibilities Boolean and Array are not used yet, but more
support is added for Integer, including propagation of bit size to
resulting target.

This commit improves error handling.

The new test category noirc_frontend_failure is added for testing input
programs that are expected to trigger an error in noirc_frontend.