Milestone: zk_dungeon
The new test program test_programs/plonky2_prove_success/zk_dungeon is a
more complicated example of a ZK program that compiles to 40Kloc of
optimized SSA in Noir.

The PLONKY2 backend is mature enough to be able to compile and prove
this program, which is the milestone.

This is reflected in the README.md.