feat: Add tracing command which auto-debugs
This commit introduces `nargo trace`. When invoked, it behaves as if
`nargo debug` is invoked, and then the `next` command is used in the
REPL until the end of the program is reached.

At the moment, just the number of `next` commands is counted and
reported.

The next commit will generate and write a `trace.json` file that will
contain a record of the execution trace.

To test:

```
cd test_programs/plonky2_prove_success/zk_dungeon/
nargo trace --trace-dir /tmp
```

It should output the following:

```
Total tracing steps: 3644
```

Note that the trace-dir is not used at the moment, so it doesn't matter
what you pass as an argument.

(cherry picked from commit 96f3b2974ef890e99862f8cee5f5d734a7f8ea61)