feat(tracing): Add scaffolding for tracing tests
This scaffolding allows the additions of tests for `nargo trace` in the
same vein as tests are added for other nargo commands.

To check locally what this PR does, checkout the repo and run

`cargo test plonky2_trace`

It should result in a single test running successfully.

To convince yourself that the test is real, try changing the info
message printed in `tooling/tracer/src/lib.rs:117` and running the test
again.

Note: there is no actual tracing logic yet, but this will be added in
following commits.