feat(tracing): return and save trace_artifact
Some of the copy-crud related to witnesses is deleted from
`trace_cmd.rs:run`, `trace_program_and_decode`, `trace_program`, and
`trace_circuit`. `trace_circuit` now returns a `TraceArtifact`, which is
saved to a file in `trace_program` and any errors are propagated back.

The test scaffold is updated to look for an `expected_trace.json` in
each test directory and compares it with the file generated by running
`nargo trace`. Also, the test verifies that `nargo trace` prints an info
message that the file has been saved.