feat(tracing): Use runtime_tracing, not TraceArtifact
This commit imports `metacraft-labs/runtime_tracing` as a dependency and
uses it instead of `TraceArtifact` to construct the tracing record.

For this reason, `tooling/nargo_cli/src/cli/fs/trace.rs` and
`tooling/noirc_artifacts/src/trace.rs` are no longer needed.

`update_record` actually propagates some information to `Tracer`, which
is now written to the output. This can be seen from the fact that
`expected_trace.json` is now non-empty.

To test, run the following:

```
cargo test plonky2_trace
```

The test still passes, exactly because of the change in
`expected_trace.json`.

**Next steps**

The next steps include recording function calls and variables.