feat(tracing): Use runtime_tracing, not TraceArtifact
This commit imports `metacraft-labs/runtime_tracing` as a dependency and
uses it instead of `TraceArtifact` to construct the tracing record.

For this reason, `tooling/nargo_cli/src/cli/fs/trace.rs` and
`tooling/noirc_artifacts/src/trace.rs` are no longer needed.

The interface of `Tracer` is a little strange. `start` needs to be used
the first time a step is recorded, and `register_step` the rest of the
time. Also, there is no way to tell that the tracer has been started,
which necessitates the introduction of `is_tracer_started`.

`update_record` actually propagates some information to `Tracer`, which
is now written to the output. This can be seen from the fact that
`expected_trace.json` is now non-empty.

**Next steps**

The next steps include recording function calls and variables.