feat(tracing): Ignore formatting in trace comparison
Before this commit, the `expected_trace.json` files had to have the
exact same format as the output of the serializer, which for size
considerations is not pretty printed.

Now, the files are not diff'ed exactly, but read into
serde Value's, which allows the comparison to ignore formatting.

This is illustrated by formatting the one test that currently exists.
The test still passes, even though the actual generated .json looks like
how `expected_trace.json` looked before this commit.