feat(trace): bugfix: track hidden steps
This has been an involved one. It turned out that `nargo debug` can skip
over some steps in the source code, if function calls are inlined, and
result in a confusing debugging experience. This is why the
`2_function_calls` test was added: it illustrates the missing steps in
the trace.

The solution is to track all locations associated with an opcode, not
just the last one. (See the two calls to `.last()` replaced with `for`
loops.) The debugger is at least smart enough to link the call sites to
the opcode, even if they are optimized away by inlining. The updated
expectation for `2_function_calls` shows that the steps for lines 6, 10,
and 11 are now included.

To see that the tests pass:

```
cargo test trace
```

Further changes:

 * this commit simplifies the testing scaffold (in
   `tooling/nargo_cli/build.rs`) because instead of comparing a message
   about the number of steps, we now compare the full traces
 * a `SourceLocation` struct is introduced to handle a file-line pair in
   a more hygienic way
 * a `DebugStepResult` is introduced, to simplify the interface between
   `trace_circuit` and `step_debugger`