feat(tracing): Add test with two files
The new test `3_two_files` contains multiple source files, where
`main.nr` calls into `foo.nr` which calls into `bar.nr`. The
`expected_trace.json` demonstrates that the tracer handles this
correctly.

Note the `path_id` fields in the JSON, which say which file the step is
in.

To run and see that the test passes:

```
cargo test trace
```