feat(tracer): Introduce TracerGlue struct
The purpose of this struct is to hold necessary state to implement the
glue between the Noir debugger and the Tracer from runtime_trace.

The first need is that sometimes a step needs to be dropped after it's
recorded. This can only be known on the next step, so state needs to be
kept.

Next commit introduces the actual state; this commit just refactors the
code into OO.