feat(tracer): Add extra step after returns
This commit adds an extra step after a return is registered, but only if
the next step would not duplicate this extra step.

This makes it clearer what's going on when one is stepping backwards
through the code, because function calls are highlighted the same way
they are on the way forwards.

(cherry picked from commit 3ce616ed5ab60f12510a15862568eee31594e5da)