refactor(plonky2): add extra arg to fn decode
An option of list of parameters was added as an extra argument to the
function `decode`. This was done so we can pass the main parameters to
`decode` and map the inputs correctly to the witness_map values.
It is wrapped with Option because `decode` is called in other places
where a list of the main parameters is not available.

Soon that will be fixed because those calls of `decode` are probably
also problematic.