feat(trace): collect plonky2 debug trace info if --trace-plonky2 is used
In case the trace command is used with the new '--trace-plonky2' option, the
compiler does a plonky2 compilation pass, which stores the generated plonky2
code (in --show-plonky2 string format), together with source line info mapping,
inside a new structure, called DebugTraceList. The new structure is then passed
as an optional parameter to the trace functions.

This is the first step towards including plonky2 information inside the trace
json file.

The changes to the trace json file format are yet to be defined.

In case of errors, during the initial plonky2 compilation, they are printed on
the screen and compilation stops, without producing a trace file. In case of
warnings, without errors, they are printed on the screen, and the trace command
continues normally with the brillig compilation. This can cause warnings to be
printed twice, because the two compilation passes are likely to produce the same
warnings. However, I'm not sure whether hiding warnings from one of the
compilation passes is any better, because it could hide diagnostic info that
could be important. In theory, the plonky2 and the brillig compilation could
produce different warnings.