Fix 2 bugs
First bug was generating an entire Verus VIR AST for every SSA
instruction. This was solved easily by stopping the recursive
translating of instructions and instead turning previous instructions
into parameters. Just like the `--show-ssa` flag does.

Second bug was related to the mismatching of the id of the parameter
which a SSA function returns. This was solved by "resolving" the id of
the value which is being returned. "Resolving" means "If
original_value_id's underlying Value has been substituted for that of
another ValueId, "resolving" will return the ValueId from which the
substitution was taken".

The solution to the second bug may seem inconsistent because only return
values ids are being resolved and all other values ids in the program are
not being resolved. As far as I have tested I have not encountered any
issues with that yet. For the future it seems like a good idea to
resolve every value id.