Fix function calling not working
Now SSA fn calls are converted properly to Verus VIR. We noticed that
argument types are not being passed during the process of convertion
from Rust to VIR which happens in Verus. That's why we reverted the
passing of argument types.

Also changed the way functions are named from "{id}" to "f{id}".