Change SSA to VIR constrains convertion logic
Made the constrains convertion the way Rust to VIR does it.
This means I added expression block wrapping and expression ghost
wrapping on top of the assert_assume expression.

I tested it and I believe it works even if we only have the ghost
wrapping and without expression block wrapping. Therefore I believe that
the expression block wrapping is not essential. I decided to leave it just
in case.