Fix unary not instruction
If we have a SSA instruction for a bit negation for a given value, how
we convert it to VIR depends on the bit width of the value. If we have a
bit width of 1 we want to convert it as a boolean negation and not as a
bit negation.