Translate Noir if clause to VIR
In Noir's SSA if expressions are flattened and removed. Instead of ifs
there are `EnableSideEffectsIf` instruction and boolean logic which is
equivalent to the if expression.
In this commit we are ignoring the `EnableSideEffectsIf` instructions
because afaik they only affect brillig functions.
We also fixed the bool to integer cast logic which SSA does so it could
represent the flattened ifs. Verus doesn't support direct bool to int
cast so I had to do it with VIR if logic.