feat(fv): Array get instructions in fv attributes
With this commit now we add the funcionality of a somewhat array get
instructions in the ensures/requires body expressions. There is a bug if
you are you using array get instructions and an assert in the body. This
bug has been filed.
What we mean by an array get instruction is `x[4]` where x is an array.