feat(fv): Finish `nargo fv` functionality
With this commit we connect all the current progress with the formal
verification to the nargo CLI. Now you can formally verify programs by
running the comman `nargo fv` in a Noir workspace.
Note! With this commit we currently do not import the Verus standard
library which is needed for verifying code which works with arrays. All
attempts to formally verify functions with arrays fails.
This will be fixed very soon when we nixify the verus-lib repo.