feat(fv): Change how we formally verify `ifs`
We had an issue where we would fail to verify correct Noir programs
because we were directly transalating the if optimiziation without
taking into consideration the instruction `enable_side_effects_if`.

Now we take into consideration the instruction enable_side_effects_if.
This way we are able to create `if` expressions which would be
semantically equivelent to the ones in Noir's AST. Because we are
generating VIR from SSA we have to reverse optimize some parts
of the code which appears in the SSA.

Also added a few tests which show that we now properly transalte if-else
statements.