feat(fv): Index arrays made of composite types
Added logic for properly indexing the flattened SSA arrays.

Noir SSA flattens arrays which are made of composite types (tuples,
structs, arrays). This resulted in having heterogeneous arrays which can
not be converted directly to Verus.

What we did to fix it is that we unflattened the array by making the
inner type a tuple. We also added a special logic for calculating the
index for the unflattened array and the index for the tuple.

Added tests which show off the new capabilities.

Co-authored-by: Kamen Maldenov <kamen.mladenov@blocksense.network>