feat(fv): Add the implication operator
For writing more readable proofs we need the implication operator.
With this commit we now add it. The user can write x ==> y where x and y
are boolean statements.

We are transforming the implication operator into !x | y which is
logically equivalent to x ==> y.

The implication operator can be used both in attributes and function's
bodies.

Added tests which showcase the new feature.