fix(fv): Add SSA pass which updates dfg's values
We now run a SSA pass which updates the values map. We do this because
we found out that the value ids for instructions inside of fv attributes
were not matching with their expected value.

This was a bug but somehow it didn't cause us any problems until the
implementation for indexing arrays of composite types.

Now that it's fixed we removed the "hack" which was circumventing the
issue inside of the `calculate_index` function.