feat(fv): Array set to VIR for simple arrays
By simple arrays we mean arrays of non composite types.

Now with this commit you can formally verify code which contains
mutation of arrays which are made of primitive types.

Soon we will also add array set support for types which are composite.

Added two tests where one of them shows the features of the current
commit and the other one the limitaitons which will fix in a new commit
soon.