feat(fv): Array Set for composite arrays
Now it's possible to verify code which contains mutation of arrays which
contain elements of a more complex type (tuples, structs, arrays).

Cargo.lock and verusStd.nix had to be updated because for this feature
we required the newest version of the verus-lib dependency.

Added some tests which showcase the newly added functionality.