fix(fv): Stop automated trait implementations from getting optimized away
Automated trait implementations were getting optimized away when they
are being used only inside of FV attributes.
With this commit we now check if they are used and if they are we don't
remove them.

Changed one of the tests which was artificially keeping those functions
by using them in the body.

Also making proofs with generic functions now seems to work. Moved the
"generics" test in "formal_verify_success" folder.