feat(fv): Mutations of arrays of nested types
With this commit we now add the feature of mutating arrays which consist
of elements of nested composite types. Previously we only supported
arrays made of simple composite types. By simple composite types we mean
tuples, arrays or structures which fields contain primitive types.

As of an example now you can mutate arrays made of tuples where the
fields of those tuples are structures, where the fields of those
structures are arrays and etc...

Added tests which show off the newly added fucntionality.