feat(fv): Add semantic analysis for quantifiers
Added type checking and name resolution for quantifier expressions.
This happens during the elaboration phase in the Noir compiler.

Because of a limitation of the Noir type system we initialize the
quantifier indexes with type `Unbounded(Any)`. Those types automatically
bound themseleves to the first type they interact with. This causes type
check failure for some expressions which in Verus' Rust won't fail.