test(fv): Add test which shows index limitations
To pass the type checking a quantifier expression must have typed
indexes. We assign the type Unbounded(Any) to the indexes.
The "problem" is that they get bounded to first type that is required of
them. E.g. a type is required when the index is used with typed
variables in an expression.
This is expected behaviour from the Noir type system but we would
have prefered if we could have assigned the Top type (a.k.a Any) to the
indexes. In this case the type wouldn't change during the elaboration
phase and it would remain Any until the monomorphization phase.

The test which is added in this commit exposes the current limitations
of the index type.