feat(fv): SSA gen for quantifier AST expression
We now transform the AST quantifier expression into Noir SSA. For each
quantifier expression we generate multiple instructions. We first
generate `allocate` instructions for the quantifier indexes. Those
allocate instructions will be ignored during the VIR gen phase.
After that we generate the `quantifier_start` instruction which serves
as an indicator where the body expression starts. We also generate a
`quantifier_end` instruction which serves as an indication where the
quantifier body ends.