feat(fv): Quantifiers to VIR
We can now use quantifiers in formal verification attributes.

Moved one of the target test programs into the success directory because
it now correctly works.
The rest of the target programs don't work because they generate SSA
instructions which are not implemented to VIR yet. Those instructions
are `allocate` and `store`.