fix(fv): Correct cast type target
When we were encountering a cast instruction with value X and target
type T we were creating a VIR expression of the kind: Cast X to
type_of(X) instead of Cast X to T. This has now been fixed.

In result one of our success tests started failing because we are doing
unsafe casts inside it. Moved it to failing tests. We will soon rewrite
the test so it doesn't use unsafe casts.