feat(fv): Handle SSA instrs `allocate` and `store`
We now properly convert the SSA instructions `allocate` and `store` to
VIR expressions. This means that we now can support more complex Noir
programs which contain arrays.

This was only one of the issues which occurred in the quantifier target
programs. There more problems that need to be fixed. That's why no tests
are being moved to the success folder.