feat(fv): Truncate referenced SSA values
We now handle SSA expression which truncate SSA values of type
reference. Before we assumed that truncation could only be possible with
values of type integer.

Added a test which shows the generation of such instructions. The test
is currently failing because Verus is not able to prove that the sum
of two numbers where the last bit is 0 the result will also have a last
bit 0.