fix(fv): Don't "if wrap" binary eq operation
When we encounter `enable_side_effects` we wrap the subsequent
dangerous instructions with an `if` expression. We were treating the
binary equality operation as a dangerous one which is incorrect. This
caused type mismatch in the VIR expressions. Added a test which
reproduces the scenario of having a `enable_side_effects` instruction
and an `eq` operation afterwards.

After applying the fix the test is successful.