fix(fv): Disable modulo division simplification
Noir optimizes the modulo division operation into bit size truncation if
the rhs of the mod operation is a power of 2. This is correct but Verus
seems to have some limitations when it comes to making conclusions
related to truncated bit size representation of numbers.

This is why we are disabling this simplification.