fix(fv): Fix incorrect load instrs in annotations
Before this fix the newly added test was generating incorrect load
instruction inside of the require annotation body.

With the added patch we now correct the load instruction to load the
proper value. This is being done during the mem2reg optimization where
store and load instructions are being optimized away. It's a convinient
place where we can do this optimization.