fix(fv): Wrong inlining when last instr is store
When the last instruction in a function is `store` we were experiencing
problems with inlining because the `load` instruction which we generally
generate was getting optimized away.

The issue was that in the Initial SSA, when we have a `store`
instruction by default it will have a `load` instruction to pair with
it. When we generate our `load` instruction we would have two `load`
instructions. This then lead to the optimization Mem2Reg to remove one
of them because it deemed it useless. After Mem2Reg we would have a
mismatch between the function return value and the `load` instruction
that remained.

The fix was to stop the generation of a `load` instruction if we already
have one.