test(fv): Add failing tests
Added a big target program which we want to formally verify. The program
is called apply_transactions. It takes accounts and transactions and it
applies each transaction to its deisgnated account. The program returns
the updated accounts and information which transactions were
successful.

Currently this target program doesn't verify successfully.

Added 2 more tests which are derived through minimization of the target
program. Each test represents a bug in our formal verification tool.