fix(fv): Update load values in annotations
During the Mem2Reg optimization pass most store and allocate instructions
are removed in the function body. In a previous fix we started updating
our load instructions in the annotations if they were referencing an
allocate variable which was assigned for removal.

We didn't take into consideration that there could be multiple store
instructions for one allocate instruction. Therefore we were randomly
chosing which store instruction to take into account. This ended up
causing issues because we need to process the closest store instruction
to the allocate instruction.

With this commit we now use the heuristic that the closest store
instruction to the allocate instruction is the one which has the
smallest parameter value.