fix(fv): Update load values in annotations
The Mem2Reg optimization pass removes most store and allocate
instructions located in the function body. In a previous fix we started
updating our load instructions in the annotations if they were
referencing an allocate variable which was assigned for removal.

We didn't take into consideration that there could be multiple store
instructions for one allocate instruction. Therefore we were randomly
choosing which store instruction to take into account. This ended up
causing issues because we need to process the closest store instruction
to the allocate instruction.

With this commit we now add timeline context where we keep information
when was each store instruction inserted. This way we can order the
store instructions and pick the earliest ones.