fix(fv): Avoid let variables in triggers
Because of our SSA to VIR approach we ended up having `let` variables
inside of triggers. Triggers are only assigned to array indexing
expression.
Sometimes the index of the array_get instruction was a `let` variable
which is problematic inside of quantifier bodies.

The solution was to recursively build the VIR expression for the index
instead of using our old approach of `let` variables.

The test which reproduces this bug was moved to directory for successful
tests.