fix(fv): Remove let array vars from triggers
In a previous fix we didn't consider the possibility of arrays being a
let defined variables. We now have handled this case.

Also previously we had missed a corner case for the recursion. If you
encounter one of the return paramaters of the function you would want
the recursion to end.
Now we add the return value ids of a function to recursion bottom set.