fix(fv): Function calls inside of ifs
Wrap with an `if` expression function calls which are located inside of
ifs. This wasn't done before and this meant that a function call was
being "executed" in both branches of the `if` expression. This for
obvious reasons is problematic.

Added tests which reproduce this scenario. Also moved the original test
for reproducing this bug to the success directory.