fix(fv): Do not wrap with `if` `lt` instructions
Before we wrapping the `lt` instruction with an `if` expression if it
was located inside of an `enable side_effects` block.

This was not needed because `lt` is not a dangerous instruction.

It caused crashes in the case of `v21 = lt v20 v5` because v21 is of
type and the default value for the else of the `if` wrapper was of a
type Int.