test(fv): Formal verification limit for Field type
Added tests which do basic arithmetic operation with a variable of type
Field. They fail to verify successfully because Verus does not support
logical reasoning for integer types with a bit size different from 8, 16,
32, 64, 128.

The default Noir Field is of bit size 254.

Added a test which shows that casting from Field to another type and
then doing the mathematical operation is verified successfully.

We have opened an issue in Verus asking if they are actually planning a
support for odd int types.
https://github.com/verus-lang/verus/issues/1503