chore(fv): Report errors when Venir has crashed
Instead of early returning during a Venir crash we now report the crash
errors using the Noir reporter.
When we were first implementing our error reporting logic we assumed that
crashes would not have spans attached to them. This assumption was
wrong. Added a test which crashes but also has span to it.

Bumped up the version of Venir in the Nix derivation because Venir has
also changed its error reporting to make this change possible.