test(fv): Binary search V2
Rewrote binary search V1 with the goal of not using an `if` expression in
the `for` loop body. Also slightly improved the function for the binary
search step in. It no longer has potential to underflow therefore the
proof is a bit shorter.

In result the binary search works with 200 element array and it takes 2
seconds to verify the program.