fix(fv): Change Field type mapping to VIR
We now translate the Field type as an VIR Int type. Because we define
the variables in exec mode there is no problem to use the mathematical
Int type.
This approach allows us to generate proofs for code which uses Field
types in arithmetic operations. Also the Field type "supports" safe
overflow behaviour because it is a ring of module p where p is prime.

The problem is that the Noir Field overflowing behaviour can not be
exactly simulated 1:1 with only using the mathematical Int type. We will
try to match that behaviour more correctly by using spec mathematical
lemmas.