feat(fv): Support wrapping overflow of Field type
This commit builds on d1498f1.

We now successfully support formally verifying code which contains
arithmetic operations with values of type Field.

This previously wasn't possible because we were not reproducing the
wrapping overflow behaviour in the VIR which we were generating.
Now we wrap all arithmetic operations with a Euclidean modulo operation
where the divisor is the Field modulus.

Added several tests which show off the new capabilities.