feat(fv): Semantic analysis for quantifiers
Added type checking and name resolution for quantifier expressions.
This happens during the elaboration phase in the Noir compiler.

Because of a limitation of the Noir type system we initialize the
quantifier indexes with type `Unbounded(Any)`. Those types automatically
bound themseleves to the first type they interact with. In result the
user sometimes has to do some extra type casting when using indexes.