feat(array-mut): Unroll deep structure modifications
Because Verus doesn't support mutation of composite structures,
we are swapping such expressions with a block of expressions which are
atomic mutaions of basic data.
Transforms unsupported nested assignments like `arr[i][j] = expr` and
`(arr[i]).1 = expr` into valid Verus-compatible operations.

We also added support for function calls to `assume()`. When we
encounter a call to a function named `assume` we transform it to the VIR
expression `AssertAssume`.