feat(fv-bridge): Typed attributes to VIR
Implemented the module which converts typed attribute expressions to
VIR.
In `fv_bridge` we are now using the new functions from
`formal_verification`. Therefore we have a completed pipeline of
parsing, type inferring and conversion to VIR.

Almost all tests which currently don't pass are tests which use Noir
structures which we don't support yet.